requirements.txt
# 데이터 처리
pandas
numpy

# 시각화
matplotlib
seaborn
plotly

# 머신러닝 (이상 탐지, 클러스터링)
scikit-learn

# 시계열 예측
prophet

# NLP / 감정 분석 (HuggingFace)
transformers
sentencepiece

# LLM (OpenAI API)
openai

# 벡터 DB (RAG)
chromadb

# 웹 앱
streamlit

# 노트북
jupyter
notebook
ipykernel

# 유틸
python-dotenv
tqdm
